// src/CodeGenerator/model/fenYeXiangYingTi.ts
/**
 * OJPlus
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { Model8 } from './model8';

export class FenYeXiangYingTi {
    /**
    * 业务消息
    */
    'message': string;
    /**
    * 业务状态码
    */
    'code': number;
    'data': Model8;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "number"
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "Model8"
        }    ];

    static getAttributeTypeMap() {
        return FenYeXiangYingTi.attributeTypeMap;
    }
}