// src/CodeGenerator/model/inlineObject1.ts
/**
 * OJPlus
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class InlineObject1 {
    /**
    * 邮箱
    */
    'email'?: string;
    /**
    * 验证码
    */
    'verification'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "verification",
            "baseName": "verification",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return InlineObject1.attributeTypeMap;
    }
}