// src/CodeGenerator/model/inlineObject10.ts
/**
 * OJPlus
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class InlineObject10 {
    /**
    * 序号
    */
    'id'?: number;
    'leaderId': number;
    'followerId': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "leaderId",
            "baseName": "leaderId",
            "type": "number"
        },
        {
            "name": "followerId",
            "baseName": "followerId",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return InlineObject10.attributeTypeMap;
    }
}