// src/CodeGenerator/model/inlineObject2.ts
/**
 * OJPlus
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class InlineObject2 {
    /**
    * 用户名，学号/工号
    */
    'username': string;
    /**
    * 密码
    */
    'password'?: string;
    /**
    * 昵称
    */
    'nickname'?: string;
    /**
    * 邮箱
    */
    'email'?: string;
    /**
    * 班级
    */
    'className'?: string;
    /**
    * 旧密码，当传入密码时必填
    */
    'oldPassword'?: string;
    /**
    * 邮箱验证码，当传入邮箱时必填
    */
    'verification'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "username",
            "baseName": "username",
            "type": "string"
        },
        {
            "name": "password",
            "baseName": "password",
            "type": "string"
        },
        {
            "name": "nickname",
            "baseName": "nickname",
            "type": "string"
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "className",
            "baseName": "className",
            "type": "string"
        },
        {
            "name": "oldPassword",
            "baseName": "oldPassword",
            "type": "string"
        },
        {
            "name": "verification",
            "baseName": "verification",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return InlineObject2.attributeTypeMap;
    }
}