// src/CodeGenerator/model/inlineObject3.ts
/**
 * OJPlus
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class InlineObject3 {
    /**
    * 账号，学号 或 邮箱
    */
    'account'?: string;
    /**
    * 密码，传入account时必填
    */
    'password'?: string;
    /**
    * 邮箱
    */
    'email'?: string;
    /**
    * 验证码，传入email时必填
    */
    'verification'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "account",
            "baseName": "account",
            "type": "string"
        },
        {
            "name": "password",
            "baseName": "password",
            "type": "string"
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "verification",
            "baseName": "verification",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return InlineObject3.attributeTypeMap;
    }
}